feat: edit & preview invitation

+++ src/app/api/member/[id]/route.ts
@@ -15,6 +15,14 @@ export const PUT = withAuth(
     { params }: { params: Promise<{ id: string }> },
   ) => {
     const { id } = await params;
+
+    if (isNaN(Number(id))) {
+      return NextResponse.json(
+        { success: false, message: 'Invalid ID' },
+        { status: 400 },
+      );
+    }
+
     const body = await request.json();
 
     const parsed = createUpdateMemberSchema
@@ -55,6 +63,14 @@ export const PATCH = withAuth(
     { params }: { params: Promise<{ id: string }> },
   ) => {
     const { id } = await params;
+
+    if (isNaN(Number(id))) {
+      return NextResponse.json(
+        { success: false, message: 'Invalid ID' },
+        { status: 400 },
+      );
+    }
+
     const body = await request.json();
 
     const parsed = updateMemberStatusSchema.safeParse(body);
@@ -93,6 +109,14 @@ export const DELETE = withAuth(
     { params }: { params: Promise<{ id: string }> },
   ) => {
     const { id } = await params;
+
+    if (isNaN(Number(id))) {
+      return NextResponse.json(
+        { success: false, message: 'Invalid ID' },
+        { status: 400 },
+      );
+    }
+
     const member = await memberService.delete(Number(id), session.user.id);
 
     return NextResponse.json({
